# backend/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Code Review Platform",
...
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/rbac/users/{user_id}": {
      "get": {
        "tags": [
          "RBAC User Management"
        ],
        "summary": "Get User",
        "description": "Get user details by ID (Admin only).\n\nRequires VIEW_USER permission (Admin only).",
        "operationId": "get_user_api_v1_rbac_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__v1__endpoints__rbac_users__UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RBAC User Management"
        ],
        "summary": "Delete User",
        "description": "Delete user (Admin only).\n\nRequires DELETE_USER permission (Admin only).\nPrevents deletion of the last admin user.",
        "operationId": "delete_user_api_v1_rbac_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rbac/audit/logs/user/{user_id}": {
      "get": {
        "tags": [
          "RBAC Audit Logs"
        ],
        "summary": "Get User Audit Logs",
        "description": "Get audit logs for a specific user (Admin only).\n\nRequires VIEW_USER permission (Admin only).\n\nQuery parameters:\n- **limit**: Maximum number of logs to return (1-1000, default 100)",
        "operationId": "get_user_audit_logs_api_v1_rbac_audit_logs_user__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of logs to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of logs to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogResponse"
                  },
                  "title": "Response Get User Audit Logs Api V1 Rbac Audit Logs User  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{user_id}/export": {
      "get": {
        "tags": [
          "User Data Management"
        ],
        "summary": "Export User Data",
        "description": "Export all user data in JSON format (GDPR compliance).\n\nValidates Requirement: 11.5 (Data export functionality in JSON format)\nValidates Requirement: 15.10 (GDPR compliance - data export)\n\nUsers can only export their own data unless they are admin.\n\n- **user_id**: UUID of the user whose data to export\n\nReturns complete user data including:\n- User profile information\n- All projects owned by the user\n- All pull requests created by the user\n- Audit logs related to the user's actions",
        "operationId": "export_user_data_api_v1_users__user_id__export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "delete": {
        "tags": [
          "User Data Management"
        ],
        "summary": "Delete User Account",
        "description": "Request account deletion (GDPR compliance - Right to be Forgotten).\n\nValidates Requirement: 11.6 (Data deletion functionality)\nValidates Requirement: 11.7 (Delete all personal data within 30 days)\nValidates Requirement: 15.10 (GDPR compliance - right to be forgotten)\n\nUsers can only delete their own account unless they are admin.\n\n- **user_id**: UUID of the user account to delete\n- **confirm_email**: Email address confirmation (must match user's email)\n- **reason**: Optional reason for account deletion\n\nImplementation:\n- Soft delete: User marked as inactive immediately\n- Scheduled deletion: Personal data deleted within 30 days\n- Audit log anonymization: User identifiers replaced with \"deleted-user-{uuid}\"\n- Projects: Transferred to system account or deleted based on policy",
        "operationId": "delete_user_account_api_v1_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountDeletionRequest": {
        "properties": {
          "confirm_email": {
            "type": "string",
            "title": "Confirm Email"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "confirm_email"
        ],
        "title": "AccountDeletionRequest",
        "description": "Account deletion request model."
      },
      "AccountDeletionResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "deletion_scheduled_at": {
            "type": "string",
            "title": "Deletion Scheduled At"
          },
          "deletion_will_complete_by": {
            "type": "string",
            "title": "Deletion Will Complete By"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "message",
          "deletion_scheduled_at",
          "deletion_will_complete_by",
          "user_id"
        ],
        "title": "AccountDeletionResponse",
        "description": "Account deletion response model."
      },
...
      "AuditLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "ip_address": {
            "type": "string",
            "title": "Ip Address"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "timestamp",
          "user_id",
          "username",
          "action",
          "resource_type",
          "resource_id",
          "ip_address",
          "user_agent",
          "success",
          "error_message"
        ],
        "title": "AuditLogResponse",
        "description": "Audit log response model."
      },
...
      "DataExportResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "projects": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Projects"
          },
          "pull_requests": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Pull Requests"
          },
          "audit_logs": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Audit Logs"
          },
          "export_timestamp": {
            "type": "string",
            "title": "Export Timestamp"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "full_name",
          "role",
          "created_at",
          "updated_at",
          "projects",
          "pull_requests",
          "audit_logs",
          "export_timestamp"
        ],
        "title": "DataExportResponse",
        "description": "User data export response model."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Generic message response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__api__v1__endpoints__rbac_users__UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "created_at",
          "updated_at",
          "is_active"
        ],
        "title": "UserResponse",
        "description": "User response model."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}